pyproject.toml
[tool.poetry]
name = "trade-flask"
version = "0.1.0"
description = ""
authors = ["vlad-terehoff <tereh_off95@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
flask-wtf = "^1.2.2"
flask-sqlalchemy = "^3.1.1"
python-dotenv = "^1.0.1"
flask-migrate = "^4.0.7"
psycopg2-binary = "^2.9.10"
pydantic-settings = "^2.7.0"
pydantic = "^2.10.4"
alembic-postgresql-enum = "^1.4.0"
flask-login = "^0.6.3"
tronpy = "^0.5.0"
flask-restx = "^1.3.0"
celery = {extras = ["redis"], version = "^5.4.0"}
flower = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
